+++ PhysicsTools/UtilAlgos/interface/CachingVariable.h
@@ -260,18 +260,6 @@ class VarSplitter : public Splitter{
   std::vector<double> slots_;
 };
 
-template <typename Object> class sortByStringFunction  {
- public:
-  sortByStringFunction(StringObjectFunction<Object> * f) : f_(f){}
-  ~sortByStringFunction(){}
-
-  bool operator() (const Object * o1, const Object * o2) {
-    return (*f_)(*o1) > (*f_)(*o2);
-  }
- private:
-  StringObjectFunction<Object> * f_;
-};
-
 template <typename Object, const char * label>
 class ExpressionVariable : public CachingVariable {
  public:
