+++ PhysicsTools/UtilAlgos/interface/StringBasedNTupler.h
@@ -25,7 +25,6 @@
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 
 #include "PhysicsTools/UtilAlgos/interface/InputTagDistributor.h"
-#include "PhysicsTools/UtilAlgos/interface/CachingVariable.h"
 
 #include "DataFormats/PatCandidates/interface/PFParticle.h"
 #include "SimDataFormats/GeneratorProducts/interface/GenEventInfoProduct.h"
